backend/requirements.txt
flask
torch
torchvision
pillow
numpy
requests
